Sidebar improvements

+++ src/main/java/gwg6784/swinggpt/gui/MainPanel.java
@@ -1,6 +1,7 @@
 package gwg6784.swinggpt.gui;
 
 import java.awt.BorderLayout;
+import java.util.List;
 
 import gwg6784.swinggpt.gui.chat.ChatPanel;
 import gwg6784.swinggpt.gui.common.Panel;
@@ -22,10 +23,10 @@ public class MainPanel extends Panel {
     public MainPanel() {
         setLayout(new BorderLayout());
 
-        this.updateSidebar();
-
         add(this.sidebarPanel, BorderLayout.WEST);
         add(this.contentSlot, BorderLayout.CENTER);
+
+        this.updateSidebar();
 
         this.conversationService.addConversationListListener(() -> this.updateSidebar());
     }
@@ -33,16 +34,24 @@ public class MainPanel extends Panel {
     private void updateSidebar() {
         this.sidebarPanel.clearItems();
 
-        this.sidebarPanel.addItem(KEY_NEW_CHAT, "New chat", () -> this.newChat());
+        this.sidebarPanel.addItem(KEY_NEW_CHAT, "+   New chat", () -> this.newChat());
+        this.sidebarPanel.addDivider();
 
-        for (Conversation conv : this.conversationService.getConversations()) {
-            this.sidebarPanel.addItem(conv.id, conv.name, () -> this.contentSlot.set(new ChatPanel(conv)));
+        List<Conversation> conversations = this.conversationService.getConversations();
+
+        for (Conversation conv : conversations) {
+            this.sidebarPanel.addItem(conv.id, conv.name,
+                    () -> this.contentSlot.set(new ChatPanel(conv)),
+                    () -> this.conversationService.deleteConversation(conv.id));
         }
 
-        this.sidebarPanel.addItem(KEY_CLEAR_CHATS, "Clear history", () -> {
-            this.conversationService.deleteAllConversations();
-            this.newChat();
-        });
+        if (!conversations.isEmpty()) {
+            this.sidebarPanel.addDivider();
+            this.sidebarPanel.addItem(KEY_CLEAR_CHATS, "Clear history", () -> {
+                this.conversationService.deleteAllConversations();
+                this.newChat();
+            });
+        }
     }
 
     private void newChat() {
